update DAGs to use shared tasks that calls tulflow
